refactor: Change ResponseMessageConstant, StatusCodeConstant Dir #22
