+++ src/main/java/sejong/user/controller/UserAuthController.java
@@ -1,5 +1,6 @@
 package sejong.user.controller;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import sejong.user.global.res.BaseResponse;
@@ -7,17 +8,14 @@ import sejong.user.global.res.DataResponse;
 import sejong.user.service.dto.UserAuthDto;
 import sejong.user.service.UserAuthService;
 
-import static sejong.user.global.res.ResponseMessageConstant.SUCCESS;
-import static sejong.user.global.res.StatusCodeConstant.OK_STATUS_CODE;
+import static sejong.user.global.res.constant.ResponseMessageConstant.SUCCESS;
+import static sejong.user.global.exception.constant.StatusCodeConstant.OK_STATUS_CODE;
 
 @RestController
 @RequestMapping("/api/user-service")
+@RequiredArgsConstructor
 public class UserAuthController {
     private final UserAuthService userAuthService;
-
-    public UserAuthController(UserAuthService userAuthService) {
-        this.userAuthService = userAuthService;
-    }
 
     @PostMapping("/login")
     public ResponseEntity<DataResponse> login(@RequestBody UserAuthDto.UserAuthRequest userAuthLoginRequest) {
